Preserve decorated pot data
