--- .github/workflows/decorated-pot-parity.yml
@@ -1,83 +1,0 @@
-name: Decorated pot parity validation
-
-on:
-  push:
-    branches: ["fix/vanilla-interaction-parity"]
-
-permissions:
-  contents: write
-
-jobs:
-  validate-and-commit:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          ref: fix/vanilla-interaction-parity
-          fetch-depth: 0
-      - uses: actions/setup-go@v5
-        with:
-          go-version-file: go.mod
-          cache: true
-      - name: Apply decorated-pot parity patch
-        run: |
-          python3 - <<'PY'
-          from pathlib import Path
-          import re
-
-          entity = Path('core/entity/entity.go')
-          source = entity.read_text()
-          if 'ItemPotDecorations' not in source:
-              source, count = re.subn(r'(\tItemDamage\s+int\n)', r'\1\tItemPotDecorations [4]string\n', source, count=1)
-              if count != 1:
-                  raise SystemExit('could not add dropped-item pot decorations')
-              entity.write_text(source)
-
-          script = Path('.github/scripts/finalize_decorated_pot_parity.py')
-          source = script.read_text()
-          needle = '''    if old not in value:\n        raise SystemExit(f"missing patch target in {path}: {old[:120]!r}")\n'''
-          replacement = '''    if old not in value:\n        if path == "core/entity/entity.go" and "ItemPotDecorations" in value:\n            return\n        raise SystemExit(f"missing patch target in {path}: {old[:120]!r}")\n'''
-          if needle not in source:
-              raise SystemExit('could not make patch helper idempotent')
-          script.write_text(source.replace(needle, replacement, 1))
-          PY
-
-          python3 .github/scripts/finalize_decorated_pot_parity.py
-
-          python3 - <<'PY'
-          from pathlib import Path
-
-          p = Path('server/decorated_pot_parity_test.go')
-          s = p.read_text()
-          s = s.replace('"GoCraft/core/game"\n', 'corentity "GoCraft/core/entity"\n\t"GoCraft/core/game"\n')
-          s = s.replace('pig := testPassiveEntity(991, "minecraft:pig")', 'pig := corentity.New(991, [16]byte{9, 9, 1}, corentity.TypePig, 0, 64, 0)')
-          p.write_text(s)
-
-          p = Path('java/handler/recipes_test.go')
-          s = p.read_text()
-          needle = '''\t\t\tcase 13:\n\t\t\t\tattributes := mustReadRecipeVarInt(t, reader)\n'''
-          replacement = '''\t\t\tcase 61: // pot_decorations\n\t\t\t\tdecorations := mustReadRecipeVarInt(t, reader)\n\t\t\t\tif decorations < 0 || decorations > 64 {\n\t\t\t\t\tt.Fatalf("invalid pot decoration count %d", decorations)\n\t\t\t\t}\n\t\t\t\tfor decoration := int32(0); decoration < decorations; decoration++ {\n\t\t\t\t\t_ = mustReadRecipeVarInt(t, reader)\n\t\t\t\t}\n\t\t\tcase 13:\n\t\t\t\tattributes := mustReadRecipeVarInt(t, reader)\n'''
-          if needle not in s:
-              raise SystemExit('could not extend recipe component parser')
-          p.write_text(s.replace(needle, replacement, 1))
-          PY
-      - name: Format
-        run: |
-          files=$(git diff --name-only -- '*.go')
-          if [ -n "$files" ]; then gofmt -w $files; fi
-      - name: Tidy
-        run: go mod tidy
-      - name: Test
-        run: go test ./...
-      - name: Vet
-        run: go vet ./...
-      - name: Build
-        run: go build ./...
-      - name: Commit validated parity fix
-        run: |
-          rm .github/scripts/finalize_decorated_pot_parity.py .github/workflows/decorated-pot-parity.yml
-          git config user.name "Elias"
-          git config user.email "86126606+el211@users.noreply.github.com"
-          git add -A
-          git commit -m "Preserve decorated pot data"
-          git push origin HEAD:fix/vanilla-interaction-parity
