+++ java/handler/block.go
@@ -146,6 +146,11 @@ func handlePlayerActionWithContext(pkt *protocol.Packet, p *player.Player, w *co
 			return nil
 		}
 		lootBlock := broken
+		potDecorations := [4]string{}
+		if broken.ResourceLocation() == "minecraft:decorated_pot" {
+			potDecorations = w.DecoratedPotDecorations(int(bx), int(by), int(bz))
+		}
+		containerItems := w.ContainerItems(int(bx), int(by), int(bz))
 		if broken.ResourceLocation() == "minecraft:decorated_pot" && held.EnchantmentLevel("minecraft:silk_touch") == 0 && blockloot.BreaksDecoratedPot(held.ItemID) {
 			lootBlock = copyBlockProperties(broken)
 			lootBlock.Properties["cracked"] = "true"
@@ -155,9 +160,10 @@ func handlePlayerActionWithContext(pkt *protocol.Packet, p *player.Player, w *co
 			enchantments[enchantment.ID] = enchantment.Level
 		}
 		lootContext := blockloot.Context{
-			Block:        lootBlock,
-			Tool:         held,
-			Enchantments: enchantments,
+			Block:          lootBlock,
+			Tool:           held,
+			Enchantments:   enchantments,
+			PotDecorations: potDecorations,
 			BlockAt: func(dx, dy, dz int) coreworld.Block {
 				return w.GetBlock(int(bx)+dx, int(by)+dy, int(bz)+dz)
 			},
@@ -183,7 +189,7 @@ func handlePlayerActionWithContext(pkt *protocol.Packet, p *player.Player, w *co
 			dropPosition := spatial.Vec3{X: float64(bx) + 0.5, Y: float64(by) + 0.5, Z: float64(bz) + 0.5}
 			ordinal := 0
 			if isJavaStorageContainer(broken.ResourceLocation()) || broken.ResourceLocation() == "minecraft:decorated_pot" || IsFurnaceContainer(broken.ResourceLocation()) {
-				for _, item := range w.ContainerItems(int(bx), int(by), int(bz)) {
+				for _, item := range containerItems {
 					if item.ItemID != "" && item.Count > 0 {
 						spawnBlockDrop(w, nextEntityID, dropPosition,
 							player.ItemStack{ItemID: item.ItemID, Count: item.Count, Damage: item.Damage}, ordinal, mgr, p.Dimension)
@@ -243,6 +249,7 @@ func spawnBlockDrop(w *coreworld.World, nextEntityID func() int32, position spat
 	entityUUID[8] = (entityUUID[8] & 0x3f) | 0x80
 	dropped := corentity.New(id, entityUUID, corentity.TypeItem, position.X, position.Y+0.25, position.Z)
 	dropped.ItemID, dropped.ItemCount, dropped.ItemDamage = stack.ItemID, stack.Count, stack.Damage
+	dropped.ItemPotDecorations = stack.PotDecorations
 	angle := float64(id+int32(ordinal)*17) * 2.399963229728653
 	dropped.VX, dropped.VY, dropped.VZ = math.Cos(angle)*0.1, 0.2, math.Sin(angle)*0.1
 	w.Entities.Add(dropped)
@@ -973,8 +980,13 @@ func handleUseItemOn(pkt *protocol.Packet, p *player.Player, w *coreworld.World,
 		applyBlockChange(px, py, pz, block, w, mgr)
 		w.SetContainerItems(px, py, pz, block.ResourceLocation(), nil)
 	case block.ResourceLocation() == "minecraft:decorated_pot":
+		block.Properties = map[string]string{
+			"facing": chestFacingFromYaw(p.Rotation.Yaw), "cracked": "false",
+			"waterlogged": strconv.FormatBool(placingInWater),
+		}
 		applyBlockChange(px, py, pz, block, w, mgr)
 		w.SetContainerItems(px, py, pz, block.ResourceLocation(), nil)
+		w.SetDecoratedPotDecorations(px, py, pz, held.NormalizedPotDecorations())
 	case block.ResourceLocation() == "minecraft:grindstone":
 		block.Properties = javaGrindstonePlacementState(face, p.Rotation.Yaw)
 		applyBlockChange(px, py, pz, block, w, mgr)
@@ -1162,7 +1174,7 @@ func useToolOrPlant(x, y, z int, face int32, target coreworld.Block, p *player.P
 				if usedSlots[slot] {
 					continue
 				}
-				items = append(items, coreworld.ContainerItem{Slot: slot, ItemID: held.ItemID, Count: 1, Damage: held.Damage, Enchantments: held.Enchantments})
+				items = append(items, coreworld.ContainerItem{Slot: slot, ItemID: held.ItemID, Count: 1, Damage: held.Damage, Enchantments: held.Enchantments, PotDecorations: held.PotDecorations})
 				w.SetContainerItems(x, y, z, target.ResourceLocation(), items)
 				if p.GameMode != player.GameModeCreative {
 					inventorySlot := player.HotbarStart + p.HeldSlot
@@ -1225,7 +1237,7 @@ func useToolOrPlant(x, y, z int, face int32, target coreworld.Block, p *player.P
 			return true
 		}
 		if stored.IsEmpty() {
-			stored = player.ItemStack{ItemID: held.ItemID, Count: 1, Damage: held.Damage, Enchantments: held.Enchantments}
+			stored = player.ItemStack{ItemID: held.ItemID, Count: 1, Damage: held.Damage, Enchantments: held.Enchantments, PotDecorations: held.PotDecorations}
 		} else {
 			stored.Count++
 		}
