+++ java/handler/chest.go
@@ -120,7 +120,7 @@ func openStorageContainer(p *player.Player, conn *network.ClientConn, w *corewor
 				continue
 			}
 			if item.Slot >= 0 && item.Slot < len(p.ContainerSlots) && item.ItemID != "" && item.Count > 0 {
-				p.ContainerSlots[item.Slot] = player.ItemStack{ItemID: item.ItemID, Count: item.Count, Damage: item.Damage, Enchantments: item.Enchantments}
+				p.ContainerSlots[item.Slot] = player.ItemStack{ItemID: item.ItemID, Count: item.Count, Damage: item.Damage, Enchantments: item.Enchantments, PotDecorations: item.PotDecorations}
 			}
 		}
 	}
@@ -153,7 +153,7 @@ func RefreshOpenStorageContainer(p *player.Player, conn *network.ClientConn, w *
 		p.ContainerSlots = make([]player.ItemStack, javaStorageContainerSize(kind))
 		for _, item := range w.ContainerItems(int(position.X), int(position.Y), int(position.Z)) {
 			if item.Slot >= 0 && item.Slot < len(p.ContainerSlots) && item.ItemID != "" && item.Count > 0 {
-				p.ContainerSlots[item.Slot] = player.ItemStack{ItemID: item.ItemID, Count: item.Count, Damage: item.Damage, Enchantments: item.Enchantments}
+				p.ContainerSlots[item.Slot] = player.ItemStack{ItemID: item.ItemID, Count: item.Count, Damage: item.Damage, Enchantments: item.Enchantments, PotDecorations: item.PotDecorations}
 			}
 		}
 		p.ContainerStateID++
@@ -189,7 +189,7 @@ func LoadChestContainerState(p *player.Player, w *coreworld.World, pos spatial.B
 	// Right half → slots 0-26.
 	for _, item := range w.ContainerItems(int(rightPos.X), int(rightPos.Y), int(rightPos.Z)) {
 		if item.Slot >= 0 && item.Slot < chestSlotCount && item.ItemID != "" && item.Count > 0 {
-			p.ContainerSlots[item.Slot] = player.ItemStack{ItemID: item.ItemID, Count: item.Count, Damage: item.Damage, Enchantments: item.Enchantments}
+			p.ContainerSlots[item.Slot] = player.ItemStack{ItemID: item.ItemID, Count: item.Count, Damage: item.Damage, Enchantments: item.Enchantments, PotDecorations: item.PotDecorations}
 		}
 	}
 	// Left half → slots 27-53.
@@ -197,7 +197,7 @@ func LoadChestContainerState(p *player.Player, w *coreworld.World, pos spatial.B
 		for _, item := range w.ContainerItems(int(leftPos.X), int(leftPos.Y), int(leftPos.Z)) {
 			slot := item.Slot + chestSlotCount
 			if item.Slot >= 0 && item.Slot < chestSlotCount && slot < slotCount && item.ItemID != "" && item.Count > 0 {
-				p.ContainerSlots[slot] = player.ItemStack{ItemID: item.ItemID, Count: item.Count, Damage: item.Damage, Enchantments: item.Enchantments}
+				p.ContainerSlots[slot] = player.ItemStack{ItemID: item.ItemID, Count: item.Count, Damage: item.Damage, Enchantments: item.Enchantments, PotDecorations: item.PotDecorations}
 			}
 		}
 	}
@@ -598,7 +598,7 @@ func persistChestContents(p *player.Player, w *coreworld.World) {
 		for slot := 0; slot < chestSlotCount && slot < len(p.ContainerSlots); slot++ {
 			stack := p.ContainerSlots[slot]
 			if !stack.IsEmpty() {
-				rightItems = append(rightItems, coreworld.ContainerItem{Slot: slot, ItemID: stack.ItemID, Count: stack.Count, Damage: stack.Damage, Enchantments: stack.Enchantments})
+				rightItems = append(rightItems, coreworld.ContainerItem{Slot: slot, ItemID: stack.ItemID, Count: stack.Count, Damage: stack.Damage, Enchantments: stack.Enchantments, PotDecorations: stack.PotDecorations})
 			}
 		}
 		w.SetContainerItems(int(pos.X), int(pos.Y), int(pos.Z), kind, rightItems)
@@ -609,7 +609,7 @@ func persistChestContents(p *player.Player, w *coreworld.World) {
 		for slot := chestSlotCount; slot < doubleChestSlotCount && slot < len(p.ContainerSlots); slot++ {
 			stack := p.ContainerSlots[slot]
 			if !stack.IsEmpty() {
-				leftItems = append(leftItems, coreworld.ContainerItem{Slot: slot - chestSlotCount, ItemID: stack.ItemID, Count: stack.Count, Damage: stack.Damage, Enchantments: stack.Enchantments})
+				leftItems = append(leftItems, coreworld.ContainerItem{Slot: slot - chestSlotCount, ItemID: stack.ItemID, Count: stack.Count, Damage: stack.Damage, Enchantments: stack.Enchantments, PotDecorations: stack.PotDecorations})
 			}
 		}
 		w.SetContainerItems(int(lp.X), int(lp.Y), int(lp.Z), kind, leftItems)
@@ -622,7 +622,7 @@ func persistChestContents(p *player.Player, w *coreworld.World) {
 		if stack.IsEmpty() {
 			continue
 		}
-		items = append(items, coreworld.ContainerItem{Slot: slot, ItemID: stack.ItemID, Count: stack.Count, Damage: stack.Damage, Enchantments: stack.Enchantments})
+		items = append(items, coreworld.ContainerItem{Slot: slot, ItemID: stack.ItemID, Count: stack.Count, Damage: stack.Damage, Enchantments: stack.Enchantments, PotDecorations: stack.PotDecorations})
 	}
 	w.SetContainerItems(int(pos.X), int(pos.Y), int(pos.Z), kind, items)
 }
