+++ server/container_automation.go
@@ -213,7 +213,7 @@ func loadAutomationContainer(world *coreworld.World, position [3]int) (string, [
 	stacks := make([]player.ItemStack, size)
 	for _, item := range world.ContainerItems(position[0], position[1], position[2]) {
 		if item.Slot >= 0 && item.Slot < size && item.ItemID != "" && item.Count > 0 {
-			stacks[item.Slot] = player.ItemStack{ItemID: item.ItemID, Count: item.Count, Damage: item.Damage, Enchantments: item.Enchantments}
+			stacks[item.Slot] = player.ItemStack{ItemID: item.ItemID, Count: item.Count, Damage: item.Damage, Enchantments: item.Enchantments, PotDecorations: item.PotDecorations}
 		}
 	}
 	return kind, stacks, true
@@ -223,7 +223,7 @@ func saveAutomationContainer(world *coreworld.World, position [3]int, kind strin
 	items := make([]coreworld.ContainerItem, 0, len(stacks))
 	for slot, stack := range stacks {
 		if !stack.IsEmpty() {
-			items = append(items, coreworld.ContainerItem{Slot: slot, ItemID: stack.ItemID, Count: stack.Count, Damage: stack.Damage, Enchantments: stack.Enchantments})
+			items = append(items, coreworld.ContainerItem{Slot: slot, ItemID: stack.ItemID, Count: stack.Count, Damage: stack.Damage, Enchantments: stack.Enchantments, PotDecorations: stack.PotDecorations})
 		}
 	}
 	world.SetContainerItems(position[0], position[1], position[2], kind, items)
@@ -289,7 +289,7 @@ func insertOneAutomationItem(kind string, destination []player.ItemStack, item p
 	}
 	for _, slot := range slots {
 		if destination[slot].IsEmpty() {
-			destination[slot] = player.ItemStack{ItemID: item.ItemID, Count: 1, Damage: item.Damage, Enchantments: item.Enchantments}
+			destination[slot] = player.ItemStack{ItemID: item.ItemID, Count: 1, Damage: item.Damage, Enchantments: item.Enchantments, PotDecorations: item.PotDecorations}
 			return true
 		}
 	}
